move hook to user dir
