+++ hooks/index.ts
@@ -2,6 +2,7 @@ export { useBalancesContext } from "./contexts/useBalancesContext";
 export { useContractsContext } from "./contexts/useContractsContext";
 export { useErrorContext } from "./contexts/useErrorContext";
 export { usePanelContext } from "./contexts/usePanelContext";
+export { useUserContext } from "./contexts/useUserContext";
 export { useVotesContext } from "./contexts/useVotesContext";
 export { useVoteTimingContext } from "./contexts/useVoteTimingContext";
 export { useWalletContext } from "./contexts/useWalletContext";
@@ -23,6 +24,8 @@ export { useOutstandingRewards } from "./queries/balances/useOutstandingRewards"
 export { useStakerDetails } from "./queries/balances/useStakerDetails";
 export { useTokenAllowance } from "./queries/balances/useTokenAllowance";
 export { useUnstakedBalance } from "./queries/balances/useUnstakedBalance";
+export { useAccountDetails } from "./queries/user/useAccountDetails";
+export { useUserVotingAndStakingDetails as useUserData } from "./queries/user/useUserVotingAndStakingDetails";
 export { useActiveVotes } from "./queries/votes/useActiveVotes";
 export { useCommittedVotes } from "./queries/votes/useCommittedVotes";
 export { useContentfulData } from "./queries/votes/useContentfulData";
@@ -31,6 +34,5 @@ export { useEncryptedVotes } from "./queries/votes/useEncryptedVotes";
 export { useHasActiveVotes } from "./queries/votes/useHasActiveVotes";
 export { usePastVotes } from "./queries/votes/usePastVotes";
 export { useRevealedVotes } from "./queries/votes/useRevealedVotes";
-export { useVoteTransactionHashes } from "./queries/votes/useVoteTransactionHashes";
 export { useUpcomingVotes } from "./queries/votes/useUpcomingVotes";
-export { useAccountDetails } from "./wallet/useAccountDetails";
+export { useVoteTransactionHashes } from "./queries/votes/useVoteTransactionHashes";
